+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Added: It's now possible to exclude locations from having any progression on them.
 
+-   Added: You can choose an arbitrary number of locations to choose randomly from for starting location.
+
 -   Changed: A Luminoth Lore scan is less likely to have hints for what was already accessible
     when that scan was found.